Add user configurable PWM max duty cycle
If the resistance of the PCB does not raise as expected when heated, or is to low when started the board will draw in excess of 5 amps and blow the fuse. Added a user configurable variable to limit the maximum PWM duty cycle thus constraining the current draw to acceptable levels.